Expanding Roster Options
This allows a single user to pick which terms to include in the rostering process. It also now has an option to include or exclude elementary options. By excluding them, you automatically move to the secondary options, which really only pulls out the TBA teachers. The additional terms are definitely more geared toward secondary buildings but may also apply to elementary buildings. However, those options are currently locked away in the secondary settings, but I'll need to update that in the next edit.